Move code around
- LED config moved from keymap to info.json
- Default RGB behavior moved from default keymap.c to tkl_f13hs.c
- VIA custom UI things moved from keymap to keyboard
