+++ keyboards/rmi_kb/tkl_f13hs/rules.mk
@@ -1,5 +1,3 @@
-RGB_MATRIX_ENABLE = yes
-RGB_MATRIX_DRIVER = WS2812
 WS2812_DRIVER = vendor
 HAPTIC_ENABLE = yes
 HAPTIC_DRIVER += SOLENOID
